Removed dead member, converted TotalLeafCount to long and sprinkled #if DEBUG where appropiate.

+++ Sharp.Core/Physic/CollisionDetection.cs
@@ -20,7 +20,7 @@ namespace Sharp.Physic
 {
 	public struct testSweepDataGetter : ISweepDataGetter
 	{
-		public int TotalLeafCount => FrustumCuller.FrozenTree.LeafCount + CollisionDetection.simulation.BroadPhase.ActiveTree.LeafCount;
+		public long TotalLeafCount => FrustumCuller.FrozenTree.LeafCount + CollisionDetection.simulation.BroadPhase.ActiveTree.LeafCount;
 
 		public ref Tree GetTree(int index)
 		{
